scale-bridge/requirements.txt
paho-mqtt
pyusb
gpiozero
rpi.gpio